pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atlas-one-step"
version = "0.1.0"
description = "ATLAS: A target-pathology atlas for one-step generation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "pandas",
  "pyyaml",
  "scikit-learn",
  "matplotlib",
  "joblib",
  "tqdm",
  "pyarrow",
  "torch",
  "torchvision",
  "h5py",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
